requirements.txt
certifi==2018.8.24
chardet==3.0.4
click==6.7
Django==1.11.15
django-cors-headers==2.4.0
django-extensions==2.1.2
django-webpack-loader==0.6.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docopt==0.6.2
Flask==0.12.2
graphviz==0.9
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pipreqs==0.4.9
pydot==1.2.4
pydotplus==2.0.2
PyJWT==1.6.4
pyparsing==2.2.1
pytz==2018.5
requests==2.19.1
six==1.11.0
typing==3.6.6
urllib3==1.23
virtualenv==15.2.0
Werkzeug==0.14.1
yarg==0.1.9
